feat(cubeAct): Add cubeActivator and cube physics
I must added codes in the player to add an impulse on the cube to move it... Also I added poor conditions in the cubeActivator because we can't detect a body in an area and disable a collisionshape at the same time (due to flush-querry). Maybe I can improve this part of code.