fix(movementGlyph): Fix selection problem
Now when we release the 'aim_with_glyph' action and after we retriggered the action, we can select directly the movableObject if we are on this. Indeed, before the fix, the cube is unselect but since it never came out from Area, the movableObject cannot reassign. So to fix that, I added an attribute 'detectedMovableObject' to handle entered and exited objects.