fix(access): Fix errs to access to a dispose objs
I have an error with the scene transition. For the GameManager and for Signal for animationPlayer. We have a problem with cannot access to a dispose object for specific instance. I fix the problem with a workaround fix. I make these instances to null before switch to a new scene. I think there is a particular behavior with the loading and unloading of scene. It's like the scenes were not destroyed so the instances of the scene were not destroyed.