Feat(menu): Create main, pause and options menu (#10)
* feat(menu): Create first main menu

There is a first main menu with some defaults : The background doesnt change the size according to the screen size. The title dont change the size according to the screen size.

* refacto(menu): Change Node types + fix bg pos

* feat(splashScreen): Add spashscreen witt animation

Now we have a splashscreen with an animation. Also I start the transitionScreen for the transitions between scenes.

* feat(transition): Add transition between scenes

* feat(transition): Add fade out for music

I added a signal when the transition start, so with this signal we can manage the music with a fade out

* feat(pause): Add pause menu

* feat(pause): Smooth transition on high layer

We had a problem with the smooth transition. PauseMenu was on the same layer that the smooth transition so the transition doesnt work well. So to fix that, I just change the layer for the sceneTransitionManager

* fix(access): Fix errs to access to a dispose objs

I have an error with the scene transition. For the GameManager and for Signal for animationPlayer. We have a problem with cannot access to a dispose object for specific instance. I fix the problem with a workaround fix. I make these instances to null before switch to a new scene. I think there is a particular behavior with the loading and unloading of scene. It's like the scenes were not destroyed so the instances of the scene were not destroyed.

* feat(options): Add first option scene

I create a simple scene but I don't have finish the scene because it's really long time to do. So I'll do that in another PR after other features

* wip(options): WIP options to change keybinding

* feat(options): First functionnal options menu

* fix(options): Try to fix problem space and left cl

* feat(options): Add options menu to other menu

* chore(menus): Clean code