Feat(lightPole): Add lightPole and movableLightPole (#11)
* feat(lightPole):Add lightPole and movableLightPole

* feat(rechargeable): Add script for rechargeable lp

Now we have 2 scripts, one for MovableLightPole and one for LightPole. Both scripts have a class that implements the interface IRechargeableLightPole. Its not a perfect solution but it's the only solution that I found. Because I can't use the same class/Script for a RigidBody and a StaticBody.

* feat(recharge): First rechargeable system

There is a problem with mask and layer for the rechargeableArea. Also this code must be applicate to the LightPole

* fix(layer): Fix problem with layer and light

* feat(recharge): Rechargeable syst for LightPole

Now we have both LightPole (movable or not) who have a rechargeable system

* feat(icon): Change game icon

* refacto(lightPole): Move methods to the helper

I use RechargeableLightPoleHelper for FadeIn and FadeOut light. Also I use it for InitializeRechargeableLightPole.

* feat(ILightPole): Add properties in ILightPole